+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Don't show user guides and Onboarding popups for users to avoid annoying.
 - Interested profiles fetch button now shown in every section of the site instead of just the "Interested in You" section.
+- Interested profiles fetch button and Photo upload button now shown in the nav bar for easier access.
 
 
 ## [1.4.2]